Added file:allocate/2
This function allows preallocation of space for files.
It succeeds only on systems (OS/filesystem) that implement this feature.
After a successful call, the actual file size might remain unchanged. This
behaviour depends on the actual file space preallocation implementation.
To guarantee the file size is updated, if desired, one needs to truncate
the file to the new length.

On GNU/Linux this function calls the Linux specific system call fallocate.
On Mac OS X it calls the fcntl F_PREALLOCATE operation.
On other Unix systems it calls the system call posix_fallocate if available.
On systems that do not implement this feature the error {error, enosys} is
returned.