Make Erlang preprocessor check dir of current file first
Make the Erlang preprocessor (epp) always check the directory of the current
source file first when looking for include files. This also applies to
includes within includes, so that any file can always refer to another
include file relative to its own position (just as in the C preprocessor).
This corrects an old partial fix in R11 that only applied to include_lib().